Minimum viable task runner
