+++ package.json
@@ -5,7 +5,7 @@
   "main": "src/index.js",
   "scripts": {
     "build": "NODE_ENV=production webpack -p --devtool=source-map",
-    "start": "babel-node server.js",
+    "start": "babel-node tasks serve",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "repository": {
